With multiple listener mode...
No success in trying to unify peer to peer and Server/Peer model :(.
It can be done but is definitely not easy: I remove the "multiple"
capability totally from now on. Handling only P2P is easier and
fits our current needs.

If "multiple"/"Server" mode is needed, it would need to be totally
re-integrated (in the hope that the impacts on the programming model
will be easier to design).