WIP: 3 problems.
- Outgoing TransportMessage must be creatable out of the blue.
The protocol can be known: it can be the highest version protocol
of the channel, but the protocol number must be set when sending.
More generally, an OutgoingMessage should be kept mutable until
sent. May be with a Freeze()/IsFrozen pattern.
- The CrisExecutorRequest is THE request object, specialized by the
Endpoint with specific data (in the channel case, the AuthToken) and
the way it relates to its IOutgoingRequest is not clear (IOutgoingRequest
should be created before the CrisExecutorRequest for fresh requests, but
cannot be for deserialized (the existing IOutgoingRequest instance
should be found in a ConcurentDictionary) and for desrialized incoming
request.
- How the "RequestBase : IOutgoingRequest" should expose its SetResult/Error
is not obvious. Part of the issue is that CrisExecutorRequest is a public
type (as it is the key generic type parameter for the executor model).