About to change my mind once again.
Pure "Groups" should be here only to propagate configurations
downstream. They don't really exist.
In "Parties", only parties should appear. This means that this may
well be the return of the "Domain", and the remove of the just
added (back) LocalParty... A Domain IS a LocalParty. It seems better
to name it "DomainParty" and its "local" impact is that it has a "$Local"
directory just like the root ApplicationIdentityService.
Can a DomainParties have no Parties?... Only the controller exists. This
seems affordable.
Can a "Domain" exists without its Controller? This would mean that this
is not a Party, only a group of party definitions: the "pure group". This is
not a Domain, only a way to factorize configurations to be DRY.

The "trick" is that the root domain is not necessarily a DomainController,
it can be ("LaToulousaine/France/$France") but can equally be an agent
("LaToulousaine/France/$Trolley1"), BUT a sub domain IS necessarily
a DomainController. This seems to be (one of) the key point I have missed.

This (eventually) seems simple (that was my intuition since I started
this but I've made so much mistakes that I doubt):
- We read the Names (Domain,Environment,Party,Full) and...
IF a PartyName is found then:
   - It is a Party.
   - If DomainName == "External", it is an ExternalParty.
   - If PartyName == DomainName.LastPart it is a DomainParty, a local
      party, the Domain Controller that can have Parties or not.
  - Otherwise it is a RemoteParty.
IF a PartyName is not found then:
  - It is "pure configuration group" that must be ignored, except that
     its ImmutableConfigurationSection is kept as a father of its Parties
     and that its inherited properties are propagated...

Once the configuration has been processed, there is no "pure groups"
at all, only 3 kind of parties: Remote, ExternalRemote and Domain.

Last question: should we continue to allow Domain in Domain or step
back to the first intuition that was that only a 2-levels configuration is
enough and simpler?
It is surely enough because the actual unique key is the party's FullName:
finding a domain "A/B" is finding the party named "A/B/$B". Once we
have this DomainParty, we can Destroy it, AddDynamicParties to it or
destroy any of its Parties (a dynamically added DomainParty must
have all its Parties also dynamic, only the initial Application configuration
should be locked).
Is it simpler? Not sure...
Keeping the recursive structure allows the "LaToulousaine" domain to
host "LaToulousaine/France" and "LaToulousaine/London" in its Parties
in the final object model. This enables the whole "LaToulousaine" to be
destroyed at once without relying on name pattern matching.
But this subordination must be done at configuration time. And this sucks:
this is not "convergent". Adding "LaToulousaine" WITH inside
"LaToulousaine/France" will not produce the same object model
as adding the former AND then the latter.
If we decide to keep the hierarchical structure, we must handle the tree,
we must move the items from where that have been defined to where
they must be exposed. This may not be a brilliant idea:
  - The final configuration is no more "visible/understandable" from
     the new position in the object tree: hidden ImmutableConfigurationSection
     are at stake that impact the feature initializations.

Note: If we want to implement a persistence layer for dynamic parties,
we MUST persist the whole ImmutableConfigurationSection "inheritance chain"
because we cannot know how the lookup of properties is done by
the Features that consume the configurations (how do you know if a given
property is "additive" - because it belongs to a set - or is "overridable"?).

Keeping the 2-levels configuration (actually reverting to the initial
intuition) seems the safest way to go but is less satisfying... This has
to be investigated and unfortunately rather early: once this behavior
is released, such a change will be a vicious breaking change (because it
is in the structure of the objects, not the API).