WIP. Adding comments and thoughts...
Forgot to mention previously that for a domain controller
name, it is the existence of the Address that determines if
we are on a RemoteParty vs. a DomainParty.

More generally "Address" and "Parties" are mutually exclusive.
To allow a DomainParty tio have no parties (that is weird), the
"Address" is the key.