WIP. MaxClockOffset and PeeringIssues.
The "Magic Nonce" idea is not dismissed: it will be possible
to support this later since this does not change the protocol:
the only change is in the fact that the Initiator will use a
Magic Nonce instead of a Nonce and that the Listener will
trust it blindly if configured. AutoTrustKey can be used for this:
if not "Never"/"Once"/"Always" then it can be a OTP string from
which a 64 bits nonce is derived.